Enables chart download functionality
Adds the ability to download charts as PNG images.

This enhancement allows users to save and share visualizations more easily. It introduces "html-to-image" and "file-saver" libraries. A download button is included in the chart header. A reference to the chart container is created to allow for proper image generation from the chart.